test(signals): add more tests
