--- packages/signals/src/computed/computed.spec.ts
@@ -1,44 +1,0 @@
-import { signal } from '../signal';
-import { computed } from './computed';
-
-describe('@signals/computed', () => {
-  test('computed has required public methods', () => {
-    const computed$ = computed(() => 0);
-
-    expect(computed$.get).toBeDefined();
-    expect(computed$.peek).toBeDefined();
-    expect(computed$.toString).toBeDefined();
-    expect(computed$.toJSON).toBeDefined();
-    expect(computed$.valueOf).toBeDefined();
-  });
-
-  test('computed gives the correct value', () => {
-    const computed$ = computed(() => 10);
-
-    expect(computed$.get()).toBe(10);
-    expect(computed$.peek()).toBe(10);
-  });
-
-  test('computed transforms to primitive value correctly', () => {
-    const computed$ = computed(() => 10);
-
-    expect(`computed is ${computed$}`).toBe('computed is 10');
-    expect(Number(computed$)).toBe(10);
-    expect(JSON.stringify(computed$)).toBe('10');
-  });
-
-  test('computed updates value with source signals correctly', () => {
-    const count1$ = signal(1);
-    const count2$ = signal(10);
-    const computed$ = computed(() => count1$.get() + count2$.get());
-
-    expect(computed$.get()).toBe(11);
-    count1$.set(2);
-    expect(computed$.get()).toBe(12);
-    expect(computed$.get()).toBe(12);
-    count1$.set(x => x + 1);
-    expect(computed$.get()).toBe(13);
-    count2$.set(20);
-    expect(computed$.get()).toBe(23);
-  });
-});
